requirements.txt
litemapy
pyvista
numpy
pillow